+++ scripts/export/entrypoint.sh
@@ -1,4 +1,4 @@
 #!/usr/bin/env bash
 set -euxo pipefail
 
-rake transpile build rename compress
+rake transpile build rename compress cleanup