+++ src/mruby_integration/models/gesture.cpp
@@ -23,11 +23,8 @@ append_models_Gesture(mrb_state* mrb)
   Gesture_class = mrb_define_class(mrb, "Gesture", mrb->object_class);
   MRB_SET_INSTANCE_TT(Gesture_class, MRB_TT_DATA);
 
-  mrb_define_class_method(mrb,
-                          Gesture_class,
-                          "enabled=",
-                          mrb_Gesture_set_enabled,
-                          MRB_ARGS_REQ(1));
+  mrb_define_class_method(
+    mrb, Gesture_class, "enabled=", mrb_Gesture_set_enabled, MRB_ARGS_REQ(1));
 
   load_ruby_models_gesture(mrb);
 }